Create a hook to simplify component function
