+++ client/my-sites/media-library/google-photos-picker-button.tsx
@@ -2,46 +2,16 @@ import { Button } from '@wordpress/components';
 import { Icon, external } from '@wordpress/icons';
 import './google-photos-picker-button.scss';
 import { useTranslate } from 'i18n-calypso';
-import { useEffect } from 'react';
 import mediaImage from 'calypso/assets/images/illustrations/media.svg';
-import { useLocalizedMoment } from 'calypso/components/localized-moment';
-import { useCreateGooglePhotosPickerSessionMutation } from 'calypso/data/media/use-google-photos-picker-session-mutation';
-import useGooglePhotosPickerSessionQuery from 'calypso/data/media/use-google-photos-picker-session-query';
-import { useDispatch, useSelector } from 'calypso/state';
-import { setPhotoPickerSession } from 'calypso/state/media/actions';
-import getGooglePhotosPickerSession from 'calypso/state/selectors/get-google-photos-picker-session';
+import useGooglePhotosPickerSession from 'calypso/my-sites/media-library/hook/use-google-photos-picker-session';
 
 const GooglePhotosPickerButton = () => {
-	const moment = useLocalizedMoment();
-	const dispatch = useDispatch();
 	const translate = useTranslate();
-	const session = useSelector( getGooglePhotosPickerSession );
-	const { mutate: createSession, isPending } = useCreateGooglePhotosPickerSessionMutation();
-	const { data: sessionData, refetch } = useGooglePhotosPickerSessionQuery( session?.id );
+	const { session, isPending } = useGooglePhotosPickerSession();
 
 	const openPickerTab = () => {
 		session?.pickerUri && window.open( session.pickerUri, '_blank' );
 	};
-
-	useEffect( () => {
-		const isSessionExpired =
-			session?.expireTime && moment( session.expireTime ).isBefore( new Date() );
-
-		if ( ! session || isSessionExpired ) {
-			createSession();
-		}
-	}, [ session, createSession, moment ] );
-
-	useEffect( () => {
-		const interval = setInterval( () => {
-			session && refetch();
-		}, 3000 );
-		return () => clearInterval( interval );
-	}, [ session, refetch ] );
-
-	useEffect( () => {
-		sessionData && dispatch( setPhotoPickerSession( sessionData ) );
-	}, [ sessionData ] );
 
 	return (
 		<div className="google-photos-picker--container media-library__connect-message">
